+++ src/sim/mcrt/parts/settings_linker.rs
@@ -11,9 +11,9 @@ use arctk_attr::load;
 #[load]
 pub struct SettingsLinker {
     /// Number of photons to simulate.
-    num_phot: u64,
+    num_phot: usize,
     /// Number of photons to simulate in each thread block.
-    block_size: u64,
+    block_size: usize,
     /// Bump distance [m].
     bump_dist: f64,
     /// Loop limit.
