+++ src/sim/render/parts/settings.rs
@@ -5,7 +5,7 @@ use crate::{access, clone, img::Gradient};
 /// General settings structure.
 pub struct Settings<'a> {
     /// Number of tracers to simulate in each thread block.
-    block_size: u64,
+    block_size: usize,
     /// Bump distance [m].
     bump_dist: f64,
     /// Loop limit.
@@ -17,7 +17,7 @@ pub struct Settings<'a> {
 }
 
 impl<'a> Settings<'a> {
-    clone!(block_size, u64);
+    clone!(block_size, usize);
     clone!(bump_dist, f64);
     clone!(loop_limit, u64);
     clone!(min_weight, f64);
@@ -27,7 +27,7 @@ impl<'a> Settings<'a> {
     #[inline]
     #[must_use]
     pub fn new(
-        block_size: u64,
+        block_size: usize,
         bump_dist: f64,
         loop_limit: u64,
         min_weight: f64,
